Export type

+++ RapidERP/RapidERP.Infrastructure/Services/ExportTypeService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using RapidERP.Application.DTOs.ExportTypeDTOs;
+using RapidERP.Application.DTOs.Shared;
 using RapidERP.Application.Interfaces;
 using RapidERP.Domain.Entities.ExportTypeModels;
 using RapidERP.Domain.Utilities;
@@ -43,11 +44,47 @@ public class ExportTypeService(RapidERPDbContext context) : IExportType
         }
     }
 
+    public async Task<RequestResponse> CreateExport(ExportDTO export)
+    {
+        try
+        {
+            ExportTypeExport masterData = new();
+            masterData.ExportTypeId = export.ExportTypeId;
+            masterData.ExportTo = export.ExportTo;
+            masterData.SourceURL = export.SourceURL;
+            
+            await context.ExportTypeExports.AddAsync(masterData);
+            await context.SaveChangesAsync();
+
+            requestResponse = new()
+            {
+                StatusCode = $"{HTTPStatusCode.Created} {HTTPStatusCode.StatusCode201}",
+                IsSuccess = true,
+                Message = ResponseMessage.CreateSuccess,
+                Data = export
+            };
+
+            return requestResponse;
+        }
+
+        catch
+        {
+            requestResponse = new()
+            {
+                StatusCode = $"{HTTPStatusCode.BadRequest} {HTTPStatusCode.StatusCode400}",
+                IsSuccess = false,
+                Message = ResponseMessage.WrongDataInput
+            };
+
+            return requestResponse;
+        }
+    }
+
     public async Task<RequestResponse> CreateSingle(ExportTypePOST masterPOST)
     {
         try
         {
-            var isExists = await context.Languages.AsNoTracking().AnyAsync(x => x.Name == masterPOST.Name);
+            var isExists = await context.ExportTypes.AsNoTracking().AnyAsync(x => x.Name == masterPOST.Name);
 
             if (isExists == false)
             {
@@ -71,8 +108,6 @@ public class ExportTypeService(RapidERPDbContext context) : IExportType
 
                 ExportTypeTracker tracker = new();
                 tracker.ExportTypeId = masterData.Id;
-                tracker.ExportTo = masterPOST.ExportTo;
-                tracker.SourceURL = masterPOST.SourceURL;
                 tracker.Browser = masterPOST.Browser;
                 tracker.Location = masterPOST.Location;
                 tracker.DeviceIP = masterPOST.DeviceIP;
@@ -80,7 +115,7 @@ public class ExportTypeService(RapidERPDbContext context) : IExportType
                 tracker.DeviceName = masterPOST.DeviceName;
                 tracker.Latitude = masterPOST.Latitude;
                 tracker.Longitude = masterPOST.Longitude;
-                tracker.ActionBy = masterPOST.ActionBy;
+                tracker.ActionBy = masterPOST.CreatedBy;
                 tracker.ActionAt = DateTime.Now;
 
                 await context.ExportTypeTrackers.AddAsync(tracker);
@@ -126,10 +161,8 @@ public class ExportTypeService(RapidERPDbContext context) : IExportType
         try
         {
             var isExists = await context.ExportTypes.AsNoTracking().AnyAsync(x => x.Id == id);
-            var isLanguageAuditExists = await context.ExportTypeAudits.AsNoTracking().AnyAsync(x => x.ExportTypeId == id);
-            var isLanguageTrackerExists = await context.ExportTypeTrackers.AsNoTracking().AnyAsync(x => x.ExportTypeId == id);
 
-            if (isExists == false || isLanguageAuditExists == false || isLanguageTrackerExists == false)
+            if (isExists == false)
             {
                 requestResponse = new()
                 {
@@ -142,7 +175,39 @@ public class ExportTypeService(RapidERPDbContext context) : IExportType
             else
             {
                 await context.ExportTypes.Where(x => x.Id == id).ExecuteDeleteAsync();
+            }
+
+            var isAuditExists = await context.ExportTypeAudits.AsNoTracking().AnyAsync(x => x.ExportTypeId == id);
+
+            if (isAuditExists == false)
+            {
+                requestResponse = new()
+                {
+                    StatusCode = $"{HTTPStatusCode.NotFound} {HTTPStatusCode.StatusCode404}",
+                    IsSuccess = false,
+                    Message = ResponseMessage.NoRecordFound
+                };
+            }
+
+            else
+            {
                 await context.ExportTypeAudits.Where(x => x.ExportTypeId == id).ExecuteDeleteAsync();
+            }
+
+            var isTrackerExists = await context.ExportTypeTrackers.AsNoTracking().AnyAsync(x => x.ExportTypeId == id);
+
+            if (isTrackerExists == false)
+            {
+                requestResponse = new()
+                {
+                    StatusCode = $"{HTTPStatusCode.NotFound} {HTTPStatusCode.StatusCode404}",
+                    IsSuccess = false,
+                    Message = ResponseMessage.NoRecordFound
+                };
+            }
+
+            else
+            {
                 await context.ExportTypeTrackers.Where(x => x.ExportTypeId == id).ExecuteDeleteAsync();
             }
 
@@ -173,11 +238,32 @@ public class ExportTypeService(RapidERPDbContext context) : IExportType
     {
         try
         {
-            var data = context.ExportTypes.AsNoTracking().AsQueryable();
+            //var data = context.ExportTypes.AsNoTracking().AsQueryable();
+            var data = (from et in context.ExportTypes
+                              join ett in context.ExportTypeTrackers on et.Id equals ett.ExportTypeId
+                              select new
+                              {
+                                  et.Id,
+                                  et.Name,
+                                  et.Description,
+                                  et.CreatedBy,
+                                  et.CreatedAt,
+                                  et.UpdatedBy,
+                                  et.UpdatedAt,
+                                  ett.Browser,
+                                  ett.Location,
+                                  ett.DeviceIP,
+                                  ett.GoogleMapUrl,
+                                  ett.DeviceName,
+                                  ett.Latitude,
+                                  ett.Longitude,
+                                  ett.ActionAt,
+                                  ett.ActionBy
+                              }).AsQueryable();
 
             if (skip == 0 || take == 0)
             {
-                var result = await data.ToListAsync();
+                var result = await data.ToListAsync() ;
 
                 requestResponse = new()
                 {
@@ -267,6 +353,56 @@ public class ExportTypeService(RapidERPDbContext context) : IExportType
         }
     }
 
+    public async Task<RequestResponse> GetAllExports(int skip, int take)
+    {
+        try
+        {
+            var data = context.ExportTypeExports
+                .Select(x => new { x.ExportTo, x.SourceURL, ExportType = x.ExportType.Name })
+                .AsNoTracking().AsQueryable();
+
+            if (skip == 0 || take == 0)
+            {
+                var result = await data.ToListAsync();
+
+                requestResponse = new()
+                {
+                    StatusCode = $"{HTTPStatusCode.OK} {HTTPStatusCode.StatusCode200}",
+                    IsSuccess = true,
+                    Message = ResponseMessage.FetchSuccess,
+                    Data = result
+                };
+            }
+
+            else
+            {
+                var result = await data.Skip(skip).Take(take).ToListAsync();
+
+                requestResponse = new()
+                {
+                    StatusCode = $"{HTTPStatusCode.OK} {HTTPStatusCode.StatusCode200}",
+                    IsSuccess = true,
+                    Message = ResponseMessage.FetchSuccessWithPagination,
+                    Data = result
+                };
+            }
+
+            return requestResponse;
+        }
+
+        catch (Exception ex)
+        {
+            requestResponse = new()
+            {
+                StatusCode = $"{HTTPStatusCode.InternalServerError} {HTTPStatusCode.StatusCode500}",
+                IsSuccess = false,
+                Message = ex.Message
+            };
+
+            return requestResponse;
+        }
+    }
+
     public async Task<RequestResponse> GetSingle(int id)
     {
         try
@@ -312,24 +448,29 @@ public class ExportTypeService(RapidERPDbContext context) : IExportType
                 .SetProperty(x => x.UpdatedBy, masterPUT.UpdatedBy)
                 .SetProperty(x => x.UpdatedAt, DateTime.Now));
 
-                await context.ExportTypeAudits.Where(x => x.Id == masterPUT.ExportTypeAuditId).ExecuteUpdateAsync(x => x
-                 .SetProperty(x => x.ExportTypeId, masterPUT.Id)
-                 .SetProperty(x => x.Description, masterPUT.Description)
-                 .SetProperty(x => x.Name, masterPUT.Name));
+                ExportTypeAudit audit = new();
+                audit.ExportTypeId = masterPUT.Id;
+                audit.Name = masterPUT.Name;
+                audit.Description = masterPUT.Description;
 
-                await context.ExportTypeTrackers.Where(x => x.Id == masterPUT.ExportTypeTrackerId).ExecuteUpdateAsync(x => x
-                 .SetProperty(x => x.ExportTypeId, masterPUT.Id)
-                 .SetProperty(x => x.ExportTo, masterPUT.ExportTo)
-                 .SetProperty(x => x.SourceURL, masterPUT.SourceURL)
-                 .SetProperty(x => x.Browser, masterPUT.Browser)
-                .SetProperty(x => x.Location, masterPUT.Location)
-                .SetProperty(x => x.DeviceIP, masterPUT.DeviceIP)
-                .SetProperty(x => x.GoogleMapUrl, masterPUT.GoogleMapUrl)
-                .SetProperty(x => x.DeviceName, masterPUT.DeviceName)
-                .SetProperty(x => x.Latitude, masterPUT.Latitude)
-                .SetProperty(x => x.Longitude, masterPUT.Longitude)
-                .SetProperty(x => x.ActionBy, masterPUT.UpdatedBy)
-                .SetProperty(x => x.ActionAt, DateTime.Now));
+                await context.ExportTypeAudits.AddAsync(audit);
+                await context.SaveChangesAsync();
+
+                 
+                ExportTypeTracker tracker = new();
+                tracker.ExportTypeId = masterPUT.Id;
+                tracker.Browser = masterPUT.Browser;
+                tracker.Location = masterPUT.Location;
+                tracker.DeviceIP = masterPUT.DeviceIP;
+                tracker.GoogleMapUrl = masterPUT.GoogleMapUrl;
+                tracker.DeviceName = masterPUT.DeviceName;
+                tracker.Latitude = masterPUT.Latitude;
+                tracker.Longitude = masterPUT.Longitude;
+                tracker.ActionBy = masterPUT.UpdatedBy;
+                tracker.ActionAt = DateTime.Now;
+
+                await context.ExportTypeTrackers.AddAsync(tracker);
+                await context.SaveChangesAsync();
 
                 requestResponse = new()
                 {
